fixes broken `make manifests`, which was also breaking `make unit`.
kustomize was interpreting the path argument
`github.com/openshift/cluster-api/config` as a URL instead of a relative file
path. It was doing something with git to retrieve files from github, put them
in a temporary directory (hence the confusing error message about a directory
in /tmp/), then failing to find what it expected.

fixes #48