complete forking of openshift version from metal3 verison
The upstream version of this repository is deprecated, in favor of
metal3-io/cluster-api-provider-metal3. The new repo only supports
newer versions of the machine-api, and so there is no point in
maintaining this repo as a "fork" any more. This patch updates the
names of imports so that we can start treating this repository as an
implementation of a different provider.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>