updates vendored cluster-api to branch release-4.5
A previous change updating controller-runtime had a side effect of causing
metrics to be turned on by default in the controller-manager, which is part of
the vendored cluster-api code. That change caused a conflict with the
machine-controller because both were deployed as containers in the same Pod,
and both tried to listen on the default metrics port of 8080.

The update introduced by this commit fixes that because the current cluster-api
code disables metrics in the controller-manager.