ensure delete removes the link between machine and host
We were not removing the host annotation from the machine, which made
it appear to always be deleting. This change reorganizes the Delete()
function in the actuator to ensure that all of the progressive steps
of deprovisioning the host and then removing both links between the
machine and host are handled.

https://bugzilla.redhat.com/show_bug.cgi?id=1855823

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>