update machine-api-operator
We need a new version of the MachineSet code to handle an issue with
the labels on Machines making them look like they do not belong to the
MachineSet. See https://bugzilla.redhat.com/show_bug.cgi?id=1857175
for details.

Pulling the new version requires a few updates to the replace
directives in go.mod. Also ran `go mod tidy`.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>