remove machine finalizer from host when deleting machine
When a Machine is deleted, any host it was using should have all
associations with Machines removed. We were removing the explicit
links to the machine, but not the finalizer that the machine
controller added. This change removes the finalizer so that hosts that
are no longer being consumed by a machine can be deleted.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>