Include Host UID in ProviderID
We currently reference the Host only by namespace and name, but this is
insufficient to uniquely identify a provider: if the BareMetalHost were
deleted and another created with the same name then the namespace+name
would match even though it could now refer to a completely different
host (and certainly not one provisioned by this Machine object).

Including the UID in the ProviderID ensures that even if the Machine
were to remain around in the Failed phase, the nodelink controller could
still link a new Machine using the replacement Host (of the same name)
to its Node as it will have a unique ProviderID.

We do not retroactively update existing Machines where the ProviderID
does not included a UID because the ProviderID of a Node is immutable,
so we would be unable to update a linked Node to match.

In theory, the ProviderID should refer to an individual deployment to a
Host, not just an individual host. However, for now the BareMetalHost
API does not provide an ID for this (in a pinch we could use the
timestamp of when provisioning completed, but it's not an intended
purpose). The mechanism of claiming the Host using the Consumer
reference should ensure that two Machines can't be associated to the
same Host object.