Get Host using ProviderID once available
Use the ProviderID to get the Host
once it is available (after the Machine reaches the Provisioned phase).

To ensure we don't mistake a different deployment to the same Host (or a
replacement Host with the same name) with our original one, only return
true from Exists() if this Machine is still set as the Host's consumer.

In the Provisioning phase (and for Machines created with an older
version of CAPBM), fall back to reading from the annotation as before.

Fixes openshift/cluster-api-provider-baremetal#104