Remove finalizer from Node on delete
Since 4dba29b397d1b242bc0593c1717a025e1f33a041 we add a finalizer on the
Node corresponding to a Machine, but we only remove it when the Node is
being deleted while the Machine is still active. Remove it when the
Machine is being deleted, otherwise the Node can never be removed.