Don't delete Machine when Host is deleted
Deleting a Host being used by a Machine is one instance of the problem
of a provider being externally deleted. Responding by deleting the
Machine object destroys the evidence, and is inconsistent with other
CAPI providers.

Instead, ensure the Machine transitions to the Failed phase if the Host
is deleted, just as we do if it is deprovisioned.

We do this by returning false from Exists() if the Host is not in the
Provisioned state. If something starts deprovisioning the Host
(including if the Host is deleted), the Machine will move to the Failed
phase.

This means we no longer need to put a finalizer on the Host to ensure we
have time to initiate the Machine delete before the Host goes away. We
still remove any existing finalizer, as it may have been added by a
previous version of CAPBM.

Fixes openshift/cluster-api-provider-baremetal#105
Fixes openshift/cluster-api-provider-baremetal#56