Add Dockerfile.rhel to match build configuration in ocp-build-data
This change adds a new Dockerfile.rhel file to control builds that
target rhel. It is inspired by the automatically generated patches which
ensure that build files match what is in the
[ocp-build-data repository](https://github.com/openshift/ocp-build-data/tree/openshift-4.6-rhel-8/images)
used for producing release artifacts.

After this change merges, the configuration files in
https://github.com/openshift/release/tree/master/ci-operator/config/openshift/cluster-api-provider-baremetal
and
https://github.com/openshift/ocp-build-data/blob/openshift-4.6/images/baremetal-machine-controller.yml
should be updated with the new dockerfile path.