add a make target to run only the unit tests
The existing unit target is useful for running tests locally by hand
because it ensures the manifests are up to date. This new target runs
only the tests, without building the manifests, and can be used in a
CI job where kustomize is not available to (a) run the tests and (b)
ensure that the output of kustomize is checked in to the repo.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>