Remove dead code from Create()
The Create() method of the actuator is only called while the Machine is
in the Provisioning phase and Exists() returns false. Currently,
Exists() will start returning true as soon as the host annotation is
present on the Machine, so there is little point doing anything in
Create() after this.

There is no way for Create() to be called after the Machine is linked to
a Node, so no point calling handleNodeFinalizer().

Calling updateMachineStatus() actually moves the Machine from the
Provisioning to Provisioned phase; were Exists() to return false in this
phase the Machine would be marked as failed, so it is safer not to call
this lest the Exists() implementation get out of sync.