Move code to release Host into a single method
Create a new method to remove both the consumer reference and our
finalizer from a Host.

This attempts to do the Right Thing if we find that this Machine no
longer owns the Host: do not clear the consumer reference if something
else is holding it, but still remove our finalizer unless there is a new
consumer that is also a Machine.