Transition Machine to Failed phase if Host deprovisioned
If something external deprovisions the Host, ensure the Machine
transitions to the Failed phase.

We do this by returning false from Exists() if the Host returns to the
Ready state. Also return false while the Host is in the Provisioning
state, so that the Machine remains in the Provisioning phase (rather
than moving to Provisioned as soon as a Host is selected) - *provided*
that the ProviderID is not set yet.