Clear Machine addresses if Host disappears
If a Host provisioned by a Machine is deleted, its IP addresses are no
longer valid as they may be assigned to a new Host. In this case, remove
the IP addresses from the Machine. This means that if a new Machine is
created on a new Host that has the same IP address, the nodelink
controller will see only the new Machine associated with the IP and will
thus be able to link it to the Node.

Two Machines sharing a ProviderID would also prevent a Node being
linked, however the incorporation of the Host UID into the ProviderID
ensures that this will change even if the new Host is recreated with the
same name as the old one.

This must be done in the Exists() method, because the Update() method is
not called again once Exists() returns false.