Add Node finalizer only during remediation
The only time it is safe to add a finalizer to a Node is when there is a
MachineHealthCheck maintaining the Machine. When a Machine is in the
failed phase, the actuator is not called again until the Machine is
deleted, so the actuator would be unable to remove the finalizer and the
Node would be immortal until the Machine is deleted. (A
MachineHealthCheck will ensure that any Machine in the failed phase will
be deleted promptly.)

At the time it was implemented it was thought potentially valuable to
always capture the Node's metadata before it was deleted for any cause,
so that it could always be restored. But in practice we only restore the
data as part of the remediation process anyway. The theoretical future
value of always capturing it before Node deletion does not outweigh the
downside of being unable to delete the Node for a failed Machine at all.