Refactor deletion of Node finalizer
Since we don't add a Node finalizer anywhere and just want to clean up
any existing ones added by a previous version of the actuator, we can do
it all in one place in Update().

This must happen before calling remediateIfNeeded(), since remediation
may delete the Node (so it would be stuck forever if the finalizer is
not removed first).